### 2024-11-09 #7429 @homerjed
Hi @crvernon,

This repository was pushed from code I wrote in preparation for a project about challenging the use of density estimation SBI cosmological inference problems.

Whilst it's only a few weeks old I have been working on this code and project for about a year now. Now seemed like a good time to push and publish since I want the readers of my paper (to be Arxiv'd in the next week or so) to run my experiments themselves. Additionally, the code is Ideal for further independent SBI analyses and beyond.

Does that cover it? Let me know, cheers :)
