### 2024-12-05 #7429 @crvernon
:wave: @jbytecode - do you think you could take this one on as editor?
