### 2024-12-05 #7429 @jbytecode
@crvernon - sure, gladly!
