### 2024-12-06 #7429 @jbytecode
@homerjed - Hi, thank you for sending your work to JOSS. I'm the handling editor of this submission.

Do you have any suggestions for potential reviewers?

The editorial bot suggests some recently published works and their corresponding authors. You can investigate them if they are similar to your work. We can invite their authors as reviewers.

You can also use the list of people (https://reviewers.joss.theoj.org/lookup) for searching suitable reviewers.

If so, please provide me a list of GitHub handles. Please don't use the `@` character for avoding unwanted notifications.

Thank you in advance.
