### 2024-12-06 #7429 @jbytecode
👋👋👋 Dear @kazewong, @mwalmsley 👋👋👋

**[SBIAX: Density-estimation simulation-based inference in JAX.](https://github.com/openjournals/joss-reviews/issues/7429)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7429).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be 23 check items for each single reviewer.  Reviewing in JOSS is well documented and is available online: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Thank you in advance!
