### 2024-12-16 #7429 @mwalmsley
Hi @jbytecode , I'm sorry but I don't think I have capacity to help with this right now (and I'm also not much of an SBI or JAX expert). The Euclid Q1 splash papers are due in Jan and that's taking a lot of my time. I hope you find someone else (I might suggest [Maya](https://scholar.google.com/citations?user=2Uya0rgAAAAJ&hl=pl)
