### 2024-12-19 #7606 @editorialbot
[REVIEW]: 'SBIAX: Density-estimation simulation-based inference in JAX'
**Submitting author:** @homerjed (<a href="http://orcid.org/0009-0002-0985-1437">Jed Homer</a>)
**Repository:** https://github.com/homerjed/sbiax
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.54
