### 2025-01-01 #7606 @jbytecode
May I have a status update for this submission, please?
