### 2025-01-03 homerjed/sbiax#1 @maja-jablonska
Fixed path stuck in shear example config
Hello! I am adding this issue while working on the review (https://github.com/openjournals/joss-reviews/issues/7606).

I noticed that when running the ```shear_nn.ipynb``` example, the ```get_shear_experiment()```, there's an error caused by a path to shear experiment data that probably got left behind in the ```configs/shear.py```. Probably the easiest way would be to add a ```data_dir``` variable to the functions outlined in that file, but maybe this example data could be included in the package as well, as naturally it might be very useful to have some starting data instantly when learning how to use the package!

Just for reference, the exact error after running the following cell:
```python
(
    alpha,                  # True parameters
    parameter_names,        # Parameter names
    mu,                     # Model expectation
    covariance,             # Data covariance
    precision,              # Inverse data covariance
    dmu,                    # Theory derivatives
    F,                      # Fisher information
    Finv,                   # Inverse Fisher information
    lower,                  # Lower prior bounds
    upper                   # Upper prior bounds
) = get_shear_experiment()
```

```
FileNotFoundError                         Traceback (most recent call last)
Cell In[4], line 12
      1 (
      2     alpha,                  # True parameters
      3     parameter_names,        # Parameter names
      4     mu,                     # Model expectation
      5     covariance,             # Data covariance
      6     precision,              # Inverse data covariance
      7     dmu,                    # Theory derivatives
      8     F,                      # Fisher information
      9     Finv,                   # Inverse Fisher information
     10     lower,                  # Lower prior bounds
     11     upper                   # Upper prior bounds
---> [1](vscode-notebook-cell:?execution_count=4&line=1)2 ) = get_shear_experiment()

File ~/Documents/sbiax/data/shear.py:1[2](vscode-notebook-cell:?execution_count=4&line=2), in get_shear_experiment()
      9 def get_shear_experiment():
     10     data_dir = "/Users/Jed.Homer/phd/sbiax/data/shear/"
---> 12     covariance   = np.loadtxt(data_dir + "covariance_cosmic_shear_PMEpaper.dat")
     1[3](vscode-notebook-cell:?execution_count=4&line=3)     precision    = np.linalg.inv(np.matrix(covariance))
     1[4](vscode-notebook-cell:?execution_count=4&line=4)     mu           = np.loadtxt(data_dir + "DES_shear-shear_a1.0_b0.[5](vscode-notebook-cell:?execution_count=4&line=5)_data_vector")[:, 1]

File ~/anaconda3/envs/sbiax/lib/python3.12/site-packages/numpy/lib/_npyio_impl.py:1395, in loadtxt(fname, dtype, comments, delimiter, converters, skiprows, usecols, unpack, ndmin, encoding, max_rows, quotechar, like)
   1392 if isinstance(delimiter, bytes):
...
    52[7](vscode-notebook-cell:?execution_count=4&line=7)                               encoding=encoding, newline=newline)
    52[8](vscode-notebook-cell:?execution_count=4&line=8) else:
--> 529     raise FileNotFoundError(f"{path} not found.")

FileNotFoundError: /Users/Jed.Homer/phd/sbiax/data/shear/covariance_cosmic_shear_PMEpaper.dat not found.
Output is truncated. View as a scrollable element or open in a text editor. Adjust cell output settings..python
```
