### 2025-01-04 2bdc33b
new toy example, review fixes, sharding
