### 2025-01-04 homerjed/sbiax#1 @homerjed
Hi Maja!

Cheers, fixed this. The data was already in the repo, but the hard coded link remained... Fixed now.

I didn't notice your pull request. Your edit was the nicer solution which I will integrate after - I've just forced the user to specify the data directory for now!

In that vain, I have included a simple toy example for non-cosmologists that runs SBI on a toy model, see `examples/toy.ipynb`.

Cheers
