+++ paper/paper.bib
@@ -151,12 +151,12 @@
       url={https://arxiv.org/abs/2202.02435}, 
 }
 
-@misc{homer,
-      title={Simulation-Based Inference Has It's Own Dodelson-Schneider Effect (but it knows it does)}, 
+@misc{homer2024simulationbasedinferencedodelsonschneidereffect,
+      title={Simulation-based inference has its own Dodelson-Schneider effect (but it knows that it does)}, 
       author={Jed Homer and Oliver Friedrich and Daniel Gruen},
       year={2024},
-      eprint={0000.00000},
+      eprint={2412.02311},
       archivePrefix={arXiv},
-      primaryClass={cs.co},
-      url={https://arxiv.org/abs/0000.00000}, 
+      primaryClass={astro-ph.CO},
+      url={https://arxiv.org/abs/2412.02311}, 
 }
