### 2025-01-04 048278f
paper adjustments
