### 2025-01-08 homerjed/sbiax#4 @kazewong
Missing references in paper
https://github.com/openjournals/joss-reviews/issues/7606

A quick round of search with `sbijax` returns these two packages that seems similar in scope and purpose:

https://github.com/jtamanas/SaxBI
https://github.com/dirmeier/sbijax

Also along the similar line, there is also a seemingly well maintained package for SBI but in torch https://github.com/sbi-dev/sbi

@homerjed I think these are relevant packages that should be cited in the paper
