### 2025-01-08 homerjed/sbiax#4 @homerjed
Cheers I have cited these to the paper.

If some context is useful (i've added a little bit alongside the references in the paper).

I began the project (and this software) before `sbijax` was developed, it also uses `haiku` which is now archived (and in my opinion is needlessly complex). `sbi` is the de-facto SBI framework but I wanted to experiment with continuous flows which it doesn't have and there's also no simple way to MCMC an ensemble of models fit with NLE. It's also quite complex for a beginner (not a dig, its an incredible project) to see what exactly is going on behind the scenes. My plan is to integrate multi-process training as well, since my paper involved running many similar experiments.

Cheers!
