+++ paper/paper.bib
@@ -160,3 +160,24 @@
       primaryClass={astro-ph.CO},
       url={https://arxiv.org/abs/2412.02311}, 
 }
+
+@article{sbimacke,
+  doi = {10.21105/joss.02505},
+  url = {https://doi.org/10.21105/joss.02505},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {52},
+  pages = {2505},
+  author = {Alvaro Tejero-Cantero and Jan Boelts and Michael Deistler and Jan-Matthis Lueckmann and Conor Durkan and Pedro J. Gonçalves and David S. Greenberg and Jakob H. Macke},
+  title = {sbi: A toolkit for simulation-based inference},
+  journal = {Journal of Open Source Software}
+}
+
+@software{sbidirmeir,
+  author = {Simon Dirmeir}
+  title = {SBIJAX: Simulation-based inference in JAX.},
+  url = {https://github.com/dirmeier/sbijax},
+  version = {0.3.0},
+  year = {2024},
+}
