+++ .github/workflows/publish.yml
@@ -8,6 +8,7 @@ on:
         - '0.0.*'
       paths:
         - sbiax/**
+        - pyproject.toml # Only if version changes!
 
 jobs:
   build:
