+++ paper/paper.bib
@@ -175,7 +175,7 @@
 }
 
 @software{sbidirmeir,
-  author = {Simon Dirmeir}
+  author = {Simon Dirmeir},
   title = {SBIJAX: Simulation-based inference in JAX.},
   url = {https://github.com/dirmeier/sbijax},
   version = {0.3.0},
