### 2025-01-08 homerjed/sbiax#5 @kazewong
Missing community guideline
https://github.com/openjournals/joss-reviews/issues/7606

I don't see any community guidelines in the current iteration of the code. Per JOSS review guideline:

Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Please add the relevant information to the repo.
