### 2025-01-08 homerjed/sbiax#6 @kazewong
Pinning packages version
https://github.com/openjournals/joss-reviews/issues/7606

In the `pyproject.toml` file, many of the dependencies do not have their version pinned except jax. This could result in installation instability down the road. I highly encourage the author to pin the version for all packages.
