### 2025-01-08 homerjed/sbiax#7 @kazewong
Adding API documentation
https://github.com/openjournals/joss-reviews/issues/7606

Currently most of the code in the repo does not come with API documentation. Please add API documentation at least with the intended function purpose and the expected input type. I think following the level of documentation as @homerjed did in https://github.com/homerjed/sbiax/blob/297317c69f89c15e799193b91eb7f402a52efb79/sbiax/compression/nn.py#L54 should be sufficient.

A side note, when using jaxtyping `Array`, I recommend to follow the example as shown in https://docs.kidger.site/jaxtyping/#example
