### 2025-01-08 0387480
jaxtyping, docstrings
