+++ sbiax/ndes/__init__.py
@@ -1,14 +1,5 @@
 from .ensemble import Ensemble
-from .multi_ensemble import MultiEnsemble
 from .cnf import CNF
 from .maf import MAF
 from .gmm import GMM
-from .ndes import (
-    nde_log_prob,
-    nde_log_prob_posterior,
-    joint_log_prob_posterior,
-    get_nde_log_prob_fn,
-    sample_nde_state,
-    sample_state
-)
 from .scaler import Scaler
