+++ sbiax/inference/nle/__init__.py
@@ -1,6 +1,1 @@
-from .log_prob_fns import (
-    get_nde_log_prob_fn,
-    sample_state,
-    _sample_initial_state
-)
 from .nuts import nuts_sample 
