+++ sbiax/meta/__init__.py
@@ -1,2 +1,1 @@
-from ._experiment import get_n_sims_values, get_nde_types, get_sbi_types
 from ._optuna import get_trial_hyperparameters, callback
