+++ pyproject.toml
@@ -28,23 +28,22 @@ urls = {repository = "https://github.com/homerjed/sbiax"}
 
 dependencies = [
     'jax==0.4.33',
-    'equinox',
-    'diffrax',
-    'optax',
-    'beartype',
-    'ml_collections',
-    'blackjax',
-    'numpy',
-    'matplotlib',
-    'einops',
-    'cloudpickle',
-    'tqdm',
-    'scikit-learn',
-    'chainconsumer',
-    'tensorflow_probability',
-    'flowjax',
-    'optuna',
-    'plotly' 
+    'equinox==0.11.8',
+    'diffrax==0.6.0',
+    'optax==0.2.3',
+    'beartype==0.19.0',
+    'ml_collections==0.1.1',
+    'blackjax==1.2.4',
+    'numpy==1.26.4',
+    'matplotlib==3.9.2',
+    'einops==0.8.0',
+    'cloudpickle==3.1.0',
+    'tqdm==4.66.5',
+    'chainconsumer==1.1.2',
+    'tensorflow_probability==0.24.0',
+    'flowjax==16.0.0',
+    'optuna==4.0.0',
+    'plotly==5.24.1' 
 ]
 
 [build-system]
