+++ sbiax/train/loader.py
@@ -1,4 +1,5 @@
-from typing import NamedTuple, Literal, Tuple
+import abc
+from typing import NamedTuple, Literal, Tuple, Generator
 import jax
 import jax.numpy as jnp
 import jax.random as jr
@@ -34,11 +35,98 @@ def sort_sample(
     )
 
 
+class _AbstractDataLoader(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def __init__(self, data, targets, *, key):
+        pass
+
+    def __iter__(self):
+        raise RuntimeError("Use `.loop` to iterate over the data loader.")
+
+    @abc.abstractmethod
+    def loop(self, batch_size):
+        pass
+
+
+class _InMemoryDataLoader(_AbstractDataLoader):
+    """
+    An in-memory data loader designed to support neural likelihood estimation (NLE) 
+    and neural posterior estimation (NPE). This loader organizes simulations and 
+    parameters into batches for training, and manages data shuffling and batching.
+
+    Attributes:
+        simulations (Array): The simulation data, with `n` samples 
+            and `x` features per sample.
+        parameters (Array): The corresponding parameter data, with `n` 
+            samples and `y` features per sample.
+        train_mode (Literal["nle", "npe"]): The training mode, either "nle" or "npe".
+        key (Key): The random key for data shuffling and reproducibility.
+
+    Methods:
+        n_batches(batch_size: int) -> int:
+            Computes the number of batches given a batch size.
+
+        loop(batch_size: int) -> Generator:
+            Generates batches of data, shuffling if necessary, while organizing 
+            data for NLE or NPE based on the training mode.
+    """
+
+    def __init__(
+        self, 
+        simulations: Float[Array, "n x"], 
+        parameters: Float[Array, "n y"], 
+        train_mode: Literal["nle", "npe"],
+        *, 
+        key: Key
+    ): 
+        self.simulations = simulations 
+        self.parameters = parameters 
+        self.train_mode = train_mode.lower()
+        self.key = key
+        assert self.train_mode.lower() in ["nle", "npe"]
+
+    @property 
+    def n_batches(self, batch_size: int) -> int:
+        return max(int(self.simulations.shape[0] / batch_size), 1)
+
+    def loop(
+        self, batch_size: int
+    ) -> Generator[Tuple[Float[Array, "batch x"], Float[Array, "batch y"]], None, None]:
+        # Loop through dataset, batching, while organising data for NPE or NLE
+        dataset_size = self.simulations.shape[0]
+        one_batch = batch_size >= dataset_size
+        key = self.key
+        indices = jnp.arange(dataset_size)
+        while True:
+            # Yield whole dataset if batch size is larger than dataset size
+            if one_batch:
+                yield sort_sample(
+                    self.train_mode, 
+                    self.simulations, 
+                    self.parameters
+                )
+            else:
+                key, subkey = jr.split(key)
+                perm = jr.permutation(subkey, indices)
+                start = 0
+                end = batch_size
+                while end < dataset_size:
+                    batch_perm = perm[start:end]
+                    yield sort_sample(
+                        self.train_mode, 
+                        self.simulations[batch_perm], 
+                        self.parameters[batch_perm] 
+                    )
+                    start = end
+                    end = start + batch_size
+
+
+
 class DataLoader(eqx.Module):
     """
         Ultra simple and jit compilable dataloader.
     """
-    arrays: tuple[Array, ...]
+    arrays: tuple[Float[Array, "n x"], Float[Array, "n y"]]
     batch_size: int
     key: Key
 
