+++ tests/test_analysis.py
@@ -72,7 +72,7 @@ def test_nle():
         train_key, 
         ensemble,
         train_mode="nle",
-        train_data=[X, Y], 
+        train_data=(X, Y), 
         opt=opt,
         n_batch=10,
         patience=2,
@@ -140,7 +140,7 @@ def test_npe():
         train_key, 
         ensemble,
         train_mode="npe",
-        train_data=[X, Y], 
+        train_data=(X, Y), 
         opt=opt,
         n_batch=10,
         patience=2,
