### 2025-01-08 homerjed/sbiax#7 @homerjed
I've removed some redundant files I forgot about, but other than that, I have added documentation strings to everything that the user will import and use themselves - e.g. `make_step` in `train.py` won't be used as its within `train()` from `train.py`. Let me know if there should be more :)

Yes the detailed `jaxtyping` was on the list, done it now for everything with `beartype` actually checking the inputs/outputs.
