### 2025-01-08 homerjed/sbiax#5 @homerjed
Missed this. Added `CONTRIBUTING.md` for those looking to contribute, thanks.
