+++ .github/workflows/tests.yml
@@ -10,7 +10,7 @@ jobs:
   run-test:
     strategy:
       matrix:
-        python-version: [ 3.12 ]
+        python-version: [ 3.10, 3.11, 3.12 ]
         os: [ ubuntu-latest ]
       fail-fast: false
 
