+++ pyproject.toml
@@ -1,9 +1,9 @@
 [project]
 name = "sbiax"
-version = "0.0.14"
+version = "0.0.15"
 description = "Fast, parallel and lightweight simulation-based inference in JAX."
 readme = "README.md"
-requires-python ="~=3.12"
+requires-python =">=3.10"
 license = {file = "LICENSE"}
 authors = [
   {name = "Jed Homer", email = "jedhmr@gmail.com"},
