+++ .github/workflows/publish.yml
@@ -20,7 +20,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version: "3.12"
+        python-version: ["3.10", "3.11", "3.12"]
     - name: Install pypa/build
       run: >-
         python3 -m
