+++ .github/workflows/publish.yml
@@ -33,9 +33,9 @@ jobs:
       run: python3 -m build
     - name: Store the distribution packages
       uses: actions/upload-artifact@v4
-      with:
-        name: python-package-distributions-${{ matrix.python-version }}
-        path: dist/
+      # with:
+      #   name: python-package-distributions-${{ matrix.python-version }}
+      #   path: dist/
 
 
 # jobs:
