+++ .github/workflows/publish.yml
@@ -1,14 +1,14 @@
-name: Publish package.
+name: Publish package
 
 on:
-    push:
-      branches:
-        - main
-      tags:
-        - '0.0.*'
-      paths:
-        - sbiax/**
-        - pyproject.toml # Only if version changes!
+  push:
+    branches:
+      - main
+    tags:
+      - '0.0.*'
+    paths:
+      - sbiax/**
+      - pyproject.toml # Only if version changes!
 
 jobs:
   build:
@@ -24,33 +24,113 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install pypa/build
-      run: >-
-        python3 -m
-        pip install
-        build
-        --user
+      run: python3 -m pip install --upgrade pip build
     - name: Build a binary wheel and a source tarball
       run: python3 -m build
     - name: Ensure dist/ exists
-      run: test -d dist/ || mkdir dist/
+      run: mkdir -p dist/
     - name: Store the distribution packages
       uses: actions/upload-artifact@v4
       with:
-        # name: python-package-distributions-${{ matrix.python-version }}
+        name: python-package-distributions-${{ matrix.python-version }}
         path: dist/
 
+  publish-to-pypi:
+    name: Publish Python 🐍 distribution 📦 to PyPI
+    if: startsWith(github.ref, 'refs/tags/')  # Only publish to PyPI on tag pushes
+    needs: build
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+    permissions:
+      id-token: write  # Mandatory for trusted publishing
+
+    steps:
+    - name: Download artifacts
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions-${{ matrix.python-version }}
+        path: dist/
+    - name: Publish distribution 📦 to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
+
+  github-release:
+    name: Create GitHub Release with Sigstore
+    needs: publish-to-pypi
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write  # Required for creating GitHub Releases
+      id-token: write  # Required for sigstore
+
+    steps:
+    - name: Download artifacts
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions-${{ matrix.python-version }}
+        path: dist/
+    - name: Sign the dists with Sigstore
+      uses: sigstore/gh-action-sigstore-python@v2.1.1
+      with:
+        inputs: |
+          ./dist/*.tar.gz
+          ./dist/*.whl
+    - name: Create GitHub Release
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      run: |
+        gh release create '${{ github.ref_name }}' --repo '${{ github.repository }}' --notes ""
+    - name: Upload artifact signatures to GitHub Release
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      run: |
+        gh release upload '${{ github.ref_name }}' dist/** --repo '${{ github.repository }}'
+
+  publish-to-testpypi:
+    name: Publish Python 🐍 distribution 📦 to TestPyPI
+    needs: build
+    runs-on: ubuntu-latest
+    environment:
+      name: testpypi
+    permissions:
+      id-token: write  # Mandatory for trusted publishing
+
+    steps:
+    - name: Download artifacts
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions-${{ matrix.python-version }}
+        path: dist/
+    - name: Publish distribution 📦 to TestPyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
+      with:
[…]
