+++ .github/workflows/publish.yml
@@ -16,7 +16,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.10", "3.11", "3.12"] # Specify multiple Python versions
+        python-version: ["3.10", "3.11", "3.12"]
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python
@@ -37,20 +37,14 @@ jobs:
 
   publish-to-pypi:
     name: Publish Python 🐍 distribution 📦 to PyPI
-    if: startsWith(github.ref, 'refs/tags/')  # Only publish to PyPI on tag pushes
+    if: startsWith(github.ref, 'refs/tags/')
     needs: build
     runs-on: ubuntu-latest
     environment:
       name: pypi
     permissions:
-      id-token: write  # Mandatory for trusted publishing
-
+      id-token: write
     steps:
-    # - name: Download artifacts
-    #   uses: actions/download-artifact@v4
-    #   with:
-    #     name: python-package-distributions-${{ matrix.python-version }}
-    #     path: dist/
     - name: Publish distribution 📦 to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
 
@@ -59,18 +53,11 @@ jobs:
     needs: publish-to-pypi
     runs-on: ubuntu-latest
     permissions:
-      contents: write  # Required for creating GitHub Releases
-      id-token: write  # Required for sigstore
-
+      contents: write
+      id-token: write
     steps:
-    # - name: Download artifacts
-    #   uses: actions/download-artifact@v4
-    #   with:
-    #     name: python-package-distributions-${{ matrix.python-version }}
-    #     path: dist/
     - name: Sign the dists with Sigstore
-      # uses: sigstore/gh-action-sigstore-python@v2.1.1
-      uses: sigstore/gh-action-sigstore-python@v3.0.0 
+      uses: sigstore/gh-action-sigstore-python@v3.0.0
       with:
         inputs: |
           ./dist/*.tar.gz
@@ -93,18 +80,127 @@ jobs:
     environment:
       name: testpypi
     permissions:
-      id-token: write  # Mandatory for trusted publishing
-
-    # steps:
-    # - name: Download artifacts
-    #   uses: actions/download-artifact@v4
-    #   with:
-    #     name: python-package-distributions-${{ matrix.python-version }}
-    #     path: dist/
+      id-token: write
+    steps:
     - name: Publish distribution 📦 to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
         repository-url: https://test.pypi.org/legacy/
+
+
+
+
+
+
+# name: Publish package
+
+# on:
+#   push:
+#     branches:
+#       - main
+#     tags:
+#       - '0.0.*'
+#     paths:
+#       - sbiax/**
+#       - pyproject.toml # Only if version changes!
+
+# jobs:
+#   build:
+#     name: Build distribution 📦
+#     runs-on: ubuntu-latest
+#     strategy:
+#       matrix:
+#         python-version: ["3.10", "3.11", "3.12"] # Specify multiple Python versions
+#     steps:
+#     - uses: actions/checkout@v4
+#     - name: Set up Python
+#       uses: actions/setup-python@v5
+#       with:
+#         python-version: ${{ matrix.python-version }}
+#     - name: Install pypa/build
+#       run: python3 -m pip install --upgrade pip build
+#     - name: Build a binary wheel and a source tarball
+#       run: python3 -m build
+#     - name: Ensure dist/ exists
+#       run: mkdir -p dist/
+#     - name: Store the distribution packages
+#       uses: actions/upload-artifact@v4
+#       with:
[…]
