+++ .github/workflows/publish.yml
@@ -16,7 +16,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.10", "3.11", "3.12"]
+        python-version: ["3.10", "3.11", "3.12"] # Specify multiple Python versions
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python
@@ -37,14 +37,19 @@ jobs:
 
   publish-to-pypi:
     name: Publish Python 🐍 distribution 📦 to PyPI
-    if: startsWith(github.ref, 'refs/tags/')
+    if: startsWith(github.ref, 'refs/tags/')  # Only publish to PyPI on tag pushes
     needs: build
     runs-on: ubuntu-latest
     environment:
       name: pypi
     permissions:
-      id-token: write
+      id-token: write  # Mandatory for trusted publishing
     steps:
+    - name: Download artifacts
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions-${{ matrix.python-version }}
+        path: dist/
     - name: Publish distribution 📦 to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
 
@@ -53,11 +58,17 @@ jobs:
     needs: publish-to-pypi
     runs-on: ubuntu-latest
     permissions:
-      contents: write
-      id-token: write
+      contents: write  # Required for creating GitHub Releases
+      id-token: write  # Required for sigstore
+
     steps:
+    - name: Download artifacts
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions-${{ matrix.python-version }}
+        path: dist/
     - name: Sign the dists with Sigstore
-      uses: sigstore/gh-action-sigstore-python@v3.0.0
+      uses: sigstore/gh-action-sigstore-python@v3.0.0 # uses: sigstore/gh-action-sigstore-python@v2.1.1
       with:
         inputs: |
           ./dist/*.tar.gz
@@ -80,16 +91,17 @@ jobs:
     environment:
       name: testpypi
     permissions:
-      id-token: write
+      id-token: write  # Mandatory for trusted publishing
     steps:
+    - name: Download artifacts
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions-${{ matrix.python-version }}
+        path: dist/
     - name: Publish distribution 📦 to TestPyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
         repository-url: https://test.pypi.org/legacy/
-
-
-
-
 
 
 # name: Publish package
@@ -110,7 +122,7 @@ jobs:
 #     runs-on: ubuntu-latest
 #     strategy:
 #       matrix:
-#         python-version: ["3.10", "3.11", "3.12"] # Specify multiple Python versions
+#         python-version: ["3.10", "3.11", "3.12"]
 #     steps:
 #     - uses: actions/checkout@v4
 #     - name: Set up Python
@@ -131,19 +143,14 @@ jobs:
 
 #   publish-to-pypi:
 #     name: Publish Python 🐍 distribution 📦 to PyPI
-#     if: startsWith(github.ref, 'refs/tags/')  # Only publish to PyPI on tag pushes
+#     if: startsWith(github.ref, 'refs/tags/')
 #     needs: build
 #     runs-on: ubuntu-latest
 #     environment:
 #       name: pypi
 #     permissions:
-#       id-token: write  # Mandatory for trusted publishing
+#       id-token: write
 #     steps:
-#     # - name: Download artifacts
-#     #   uses: actions/download-artifact@v4
-#     #   with:
-#     #     name: python-package-distributions-${{ matrix.python-version }}
-#     #     path: dist/
[…]
