+++ .github/workflows/publish.yml
@@ -1,107 +1,348 @@
-name: Publish package
-
-on:
-  push:
-    branches:
-      - main
-    tags:
-      - '0.0.*'
-    paths:
-      - sbiax/**
-      - pyproject.toml # Only if version changes!
-
+name: Publish Python 🐍 distribution 📦 to PyPI and TestPyPI
+on: push
 jobs:
   build:
     name: Build distribution 📦
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.10", "3.11", "3.12"] # Specify multiple Python versions
+        python-version: ["3.10", "3.11", "3.12"]
     steps:
     - uses: actions/checkout@v4
-    - name: Set up Python
+      with:
+        persist-credentials: false
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install pypa/build
-      run: python3 -m pip install --upgrade pip build
+      run: >-
+        python3 -m
+        pip install
+        build
+        --user
     - name: Build a binary wheel and a source tarball
       run: python3 -m build
-    - name: Ensure dist/ exists
-      run: mkdir -p dist/
     - name: Store the distribution packages
       uses: actions/upload-artifact@v4
       with:
-        name: python-package-distributions-${{ matrix.python-version }}
+        name: python-package-distributions
         path: dist/
 
   publish-to-pypi:
-    name: Publish Python 🐍 distribution 📦 to PyPI
-    if: startsWith(github.ref, 'refs/tags/')  # Only publish to PyPI on tag pushes
-    needs: build
+    name: >-
+      Publish Python 🐍 distribution 📦 to PyPI
+    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
+    needs:
+    - build
     runs-on: ubuntu-latest
     environment:
       name: pypi
+      url: https://pypi.org/p/sbiax  # Replace <package-name> with your PyPI project name
     permissions:
-      id-token: write  # Mandatory for trusted publishing
+      id-token: write  # IMPORTANT: mandatory for trusted publishing
     steps:
-    - name: Download artifacts
+    - name: Download all the dists
       uses: actions/download-artifact@v4
       with:
-        name: python-package-distributions-${{ matrix.python-version }}
+        name: python-package-distributions
         path: dist/
     - name: Publish distribution 📦 to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
 
   github-release:
-    name: Create GitHub Release with Sigstore
-    needs: publish-to-pypi
+    name: >-
+      Sign the Python 🐍 distribution 📦 with Sigstore
+      and upload them to GitHub Release
+    needs:
+    - publish-to-pypi
     runs-on: ubuntu-latest
     permissions:
-      contents: write  # Required for creating GitHub Releases
-      id-token: write  # Required for sigstore
-
+      contents: write  # IMPORTANT: mandatory for making GitHub Releases
+      id-token: write  # IMPORTANT: mandatory for sigstore
     steps:
-    - name: Download artifacts
+    - name: Download all the dists
       uses: actions/download-artifact@v4
       with:
-        name: python-package-distributions-${{ matrix.python-version }}
+        name: python-package-distributions
         path: dist/
     - name: Sign the dists with Sigstore
-      uses: sigstore/gh-action-sigstore-python@v3.0.0 # uses: sigstore/gh-action-sigstore-python@v2.1.1
[…]
