+++ .github/workflows/publish.yml
@@ -1,92 +1,118 @@
-name: Publish Python 🐍 distribution 📦 to PyPI and TestPyPI
-on: push
+name: Publish 📡 Python 🐍 distributions 📦
+
+on:
+  push:
+    tags:
+      - 0.0.*
+    paths:
+      - sbiax/**
+      - pyproject.toml # Only if version changes!
+
 jobs:
-  build:
-    name: Build distribution 📦
+  # lint:
+  #   name: Check 🧐 Python 🐍 package 🗂️
+  #   runs-on: ubuntu-latest
+
+  test:
+    name: Test 🔍 package 🗂️
     runs-on: ubuntu-latest
     strategy:
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
+
     steps:
-    - uses: actions/checkout@v4
-      with:
-        persist-credentials: false
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install pypa/build
-      run: >-
-        python3 -m
-        pip install
-        build
-        --user
-    - name: Build a binary wheel and a source tarball
-      run: python3 -m build
-    - name: Store the distribution packages
-      uses: actions/upload-artifact@v4
-      with:
-        name: python-package-distributions
-        path: dist/
+      - uses: actions/checkout@v4
+
+      - name: Setup 🛠️ Python 🐍
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install 🛠️ dependencies 📚
+        run: |
+          pip install --upgrade pip
+          pip install --requirement=requirements.txt --requirement=tests/requirements.txt
+
+      - name: Run ⏱️ tests 📊
+        run: pytest
+
+  build:
+    name: Build 🏗️ distributions 📦
+    needs: [test]
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Setup 🛠️ Python 🐍
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.x"
+
+      - name: Build 🏗️ a source distribution 🗃️ and a binary wheel 🛞
+        run: pipx run build --outdir=distributions
+
+      - name: Upload 📤 the built distributions 📦
+        uses: actions/upload-artifact@v4
+        with:
+          name: distributions-py${{ matrix.python-version }} # name: distributions
+          path: distributions/
 
   publish-to-pypi:
-    name: >-
-      Publish Python 🐍 distribution 📦 to PyPI
-    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
-    needs:
-    - build
+    name: Publish 📡 Python 🐍 distributions 📦 to PyPI 🌐
+    needs: [build]
     runs-on: ubuntu-latest
     environment:
-      name: pypi
-      url: https://pypi.org/p/sbiax  # Replace <package-name> with your PyPI project name
+      name: release
+      url: https://pypi.org/p/sbiax
     permissions:
-      id-token: write  # IMPORTANT: mandatory for trusted publishing
-    steps:
-    - name: Download all the dists
-      uses: actions/download-artifact@v4
-      with:
-        name: python-package-distributions
-        path: dist/
-    - name: Publish distribution 📦 to PyPI
-      uses: pypa/gh-action-pypi-publish@release/v1
+      id-token: write
 
-  github-release:
-    name: >-
-      Sign the Python 🐍 distribution 📦 with Sigstore
-      and upload them to GitHub Release
-    needs:
-    - publish-to-pypi
+    steps:
+      - name: Download 📥 all the distributions 📦
[…]
