+++ .github/workflows/publish.yml
@@ -32,7 +32,7 @@ jobs:
         run: |
           pip install --upgrade pip
           pip install -e .
-        # pip install --requirement=requirements.txt --requirement=tests/requirements.txt
+          pip install --requirement=tests/requirements.txt
 
       - name: Run ⏱️ tests 📊
         run: pytest
