+++ .github/workflows/publish.yml
@@ -9,10 +9,6 @@ on:
       - pyproject.toml # Only if version changes!
 
 jobs:
-  # lint:
-  #   name: Check 🧐 Python 🐍 package 🗂️
-  #   runs-on: ubuntu-latest
-
   test:
     name: Test 🔍 package 🗂️
     runs-on: ubuntu-latest
@@ -48,7 +44,7 @@ jobs:
       - name: Setup 🛠️ Python 🐍
         uses: actions/setup-python@v5
         with:
-          python-version: "3.x"
+          python-version: "3.12"
 
       - name: Build 🏗️ a source distribution 🗃️ and a binary wheel 🛞
         run: pipx run build --outdir=distributions
@@ -61,6 +57,7 @@ jobs:
 
   publish-to-pypi:
     name: Publish 📡 Python 🐍 distributions 📦 to PyPI 🌐
+    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
     needs: [build]
     runs-on: ubuntu-latest
     environment:
@@ -147,7 +144,7 @@ jobs:
     - name: Download all the dists
       uses: actions/download-artifact@v4
       with:
-        pattern: python-package-distributions-*
+        pattern: distributions-* #python-package-distributions-*
         path: distributions/  
         merge-multiple: true
     - name: Publish distribution 📦 to TestPyPI
