+++ .github/workflows/publish.yml
@@ -65,7 +65,6 @@ jobs:
       url: https://pypi.org/p/sbiax
     permissions:
       id-token: write
-
     steps:
       - name: Download 📥 all the distributions 📦
         uses: actions/download-artifact@v4
@@ -86,7 +85,6 @@ jobs:
     permissions:
       id-token: write
       contents: write
-
     steps:
       - name: Download 📥 all the distributions 📦
         uses: actions/download-artifact@v4
@@ -94,13 +92,12 @@ jobs:
           pattern: distributions
           path: distributions/
           merge-multiple: true
-
       - name: Sign 🔑 the distributions 📦 with Sigstore
         uses: sigstore/gh-action-sigstore-python@v3.0.0
         with:
-          inputs: |
-            distributions/*.tar.gz
-            distributions/*.whl
+          inputs: >-
+            ./distributions/*.tar.gz
+            ./distributions/*.whl
       - name: Create GitHub Release
         env:
           GITHUB_TOKEN: ${{ github.token }}
