+++ .github/workflows/publish.yml
@@ -1,143 +1,281 @@
-name: Publish 📡 Python 🐍 distributions 📦
+name: Publish package.
 
 on:
   push:
+    branches:
+      - main
     tags:
-      - 0.0.*
+      - '0.0.*'
     paths:
       - sbiax/**
-      - pyproject.toml # Only if version changes!
+      - pyproject.toml
 
 jobs:
-  test:
-    name: Test 🔍 package 🗂️
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.10", "3.11", "3.12"]
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Setup 🛠️ Python 🐍
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Install 🛠️ dependencies 📚
-        run: |
-          pip install --upgrade pip
-          pip install -e .
-          pip install --requirement=tests/requirements.txt
-
-      - name: Run ⏱️ tests 📊
-        run: pytest
-
   build:
-    name: Build 🏗️ distributions 📦
-    needs: [test]
+    name: Build distribution 📦
     runs-on: ubuntu-latest
 
     steps:
-      - uses: actions/checkout@v4
+    - uses: actions/checkout@v4
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      strategy:
+        matrix:
+          python-version: ["3.10", "3.11", "3.12"]
 
-      - name: Setup 🛠️ Python 🐍
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.12"
-
-      - name: Build 🏗️ a source distribution 🗃️ and a binary wheel 🛞
-        run: pipx run build --outdir=distributions
-
-      - name: Upload 📤 the built distributions 📦
-        uses: actions/upload-artifact@v4
-        with:
-          name: distributions-py${{ matrix.python-version }} # name: distributions
-          path: distributions/
+    - name: Install pypa/build
+      run: >-
+        python3 -m
+        pip install
+        build
+        --user
+    - name: Build a binary wheel and a source tarball
+      run: python3 -m build
+    - name: Store the distribution packages
+      uses: actions/upload-artifact@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+        name: python-package-distributions
+        path: dist-py${{ matrix.python-version }}/
 
   publish-to-pypi:
-    name: Publish 📡 Python 🐍 distributions 📦 to PyPI 🌐
+    name: >-
+      Publish Python 🐍 distribution 📦 to PyPI
     if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
-    needs: [build]
+    needs:
+    - build
     runs-on: ubuntu-latest
     environment:
-      name: release
-      url: https://pypi.org/p/sbiax
+      name: pypi
+      url: https://pypi.org/p/sbiax  # Replace <package-name> with your PyPI project name
     permissions:
-      id-token: write
+      id-token: write  # IMPORTANT: mandatory for trusted publishing
+
     steps:
-      - name: Download 📥 all the distributions 📦
-        uses: actions/download-artifact@v4
-        with:
-          pattern: distributions-* #pattern: distributions
-          path: distributions/
-          merge-multiple: true
+    - name: Download all the dists
+      uses: actions/download-artifact@v4
+      with:
+        pattern: dist-* # !!
+        name: python-package-distributions
+        path: dist/
+        merge-multiple: true
+    - name: Publish distribution 📦 to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
 
-      - name: Publish 📡 distributions 📦 to PyPI 🌐
[…]
