+++ .github/workflows/publish.yml
@@ -11,6 +11,28 @@ on:
       - pyproject.toml
 
 jobs:
+  test:
+    name: Test 🔍 package 🗂️
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup 🛠️ Python 🐍
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install 🛠️ dependencies 📚
+        run: |
+          pip install --upgrade pip
+          pip install -e .
+          pip install --requirement=tests/requirements.txt
+
+      - name: Run ⏱️ tests 📊
+        run: pytest
+
   build:
     name: Build distribution 📦
     runs-on: ubuntu-latest
@@ -19,9 +41,8 @@ jobs:
     - uses: actions/checkout@v4
     - name: Set up Python
       uses: actions/setup-python@v5
-      strategy:
-        matrix:
-          python-version: ["3.10", "3.11", "3.12"]
+      with:
+        python-version: "3.12"
 
     - name: Install pypa/build
       run: >-
@@ -34,9 +55,11 @@ jobs:
     - name: Store the distribution packages
       uses: actions/upload-artifact@v4
       with:
-        python-version: ${{ matrix.python-version }}
         name: python-package-distributions
-        path: dist-py${{ matrix.python-version }}/
+        path: dist/
+        # python-version: ${{ matrix.python-version }}
+        # name: python-package-distributions
+        # path: dist-py${{ matrix.python-version }}/
 
   publish-to-pypi:
     name: >-
