### 2025-01-10 homerjed/sbiax#4 @kazewong
Sounds good. JOSS review permits submission of new package even if there are similar solutions available, the only requirement is to acknowledge the existence of other solutions. This is now completed and can be closed
