### 2025-01-10 homerjed/sbiax#7 @kazewong
I think it will be nice if you can add documentation to functions and classes which the users are not necessarily expected to import and use. This could help potential contributors to make changes if they need to. For example,

https://github.com/homerjed/sbiax/blob/4653abed1c996d454fad181acf869b2cea76ba66/sbiax/inference/nle/nuts.py#L98
https://github.com/homerjed/sbiax/blob/4653abed1c996d454fad181acf869b2cea76ba66/sbiax/compression/linear.py#L15
