### 2025-01-10 3641ee1
docs for linear compression
