+++ sbiax/compression/linear.py
@@ -1,33 +1,93 @@
 import jax.numpy as jnp
 import jax.random as jr 
-
-"""
-    Tools for linear compression.
-    - derive an MLE given a datavector and the data covariance, expectation,
-      theory derivatives and parameters.
-"""
+from jaxtyping import Key, Array, Float, jaxtyped
+from beartype import beartype as typechecker
 
 
-def linearized_model(_alpha, mu, alpha, derivatives):
-    return mu + jnp.dot(_alpha - alpha, derivatives)
-
-
-def simulator(key, parameters, alpha, mu, derivatives, covariance):
+@jaxtyped(typechecker=typechecker)
+def linearized_model(
+    pi_: Float[Array, "p"], 
+    mu: Float[Array, "d"], 
+    pi: Float[Array, "p"], 
+    dmu: Float[Array, "p d"]
+) -> Float[Array, "p"]:
     """
-        Simulate from a Gaussian likelihood.
+        Calculate a linearised model prediction given a set of parameters.
+
+        Args:
+            pi_ (`Array`): The parameters of the model to calculate.
+            mu (`Array`): The model evaluated at the estimated set of parameters `pi`.
+            pi (`Array`): A fiducial set of parameters (e.g. those the `covariance` is calculated at)
+            dmu (`Array`): The first-order theory derivatives (for the implicitly assumed linear model, 
+                these are parameter independent!)
+
+        Returns:
+            `Array`: a model prediction at the parameters `pi_`
+    """
+    return mu + jnp.dot(pi_ - pi, dmu)
+
+
+@jaxtyped(typechecker=typechecker)
+def simulator(
+    key: Key[jnp.ndarray, "..."], 
+    pi_: Float[Array, "p"], 
+    pi: Float[Array, "p"], 
+    mu: Float[Array, "d"], 
+    dmu: Float[Array, "p d"], 
+    covariance: Float[Array, "d d"]
+) -> Float[Array, "d"]:
+    """
+        Simulate from a Gaussian likelihood defined by a model `mu` and a data covariance `covariance`.
+
+        Args:
+            key (`Key`): A JAX random key.
+            pi (`Array`): The estimated parameters of the datavector.
+            alpha (`Array`): A fiducial set of parameters (e.g. those the `covariance` is calculated at)
+            mu (`Array`): The model evaluated at the estimated set of parameters `pi`.
+            dmu (`Array`): The first-order theory derivatives (for the implicitly assumed linear model, 
+                these are parameter independent!)
+            covariance (`Array`): The data covariance matrix.
+
+        Returns:
+            `Array`: a datavector drawn from the Gaussian likelihood.
     """
     d = jr.multivariate_normal(
         key=key, 
         mean=linearized_model(
-            _alpha=parameters, 
+            pi_=pi_, 
             mu=mu, 
-            alpha=alpha, 
-            derivatives=derivatives
+            pi=pi, 
+            dmu=dmu 
         ),
         cov=covariance
     ) 
     return d
 
 
-def mle(d, pi, Finv, mu, dmu, precision):
+@jaxtyped(typechecker=typechecker)
+def mle(
+    d: Float[Array, "d"], 
+    pi: Float[Array, "p"], 
+    Finv: Float[Array, "p p"], 
+    mu: Float[Array, "d"], 
+    dmu: Float[Array, "p d"], 
+    precision: Float[Array, "d d"]
+) -> Float[Array, "p"]:
+    """
+        Calculates a maximum likelihood estimator (MLE) from a datavector by
+        assuming a linear model `mu` in parameters `pi` and using
+
+        Args:
+            d (`Array`): The datavector to compress.
+            p (`Array`): The estimated parameters of the datavector (e.g. a fiducial set).
[…]
