### 2025-01-10 d193368
docstrings for NUTS sampling
