### 2025-01-10 1f8b63b
loader loop method docstring
