### 2025-01-10 3522ec3
removed redundant npe/inference, docstrings etc
