+++ sbiax/inference/npe/__init__.py
@@ -1,1 +1,0 @@
-from .npe import sample_nde, posterior_log_prob
