+++ sbiax/inference/__init__.py
@@ -1,2 +1,1 @@
-from .npe import sample_nde
-from .nle import nuts_sample
+from .nuts import nuts_sample
