+++ sbiax/inference/nle/__init__.py
@@ -1,1 +1,0 @@
-from .nuts import nuts_sample 
