### 2025-01-10 04d40b2
test fitting nn, docstrings
