+++ sbiax/train/train.py
@@ -27,6 +27,23 @@ def shard_batch(
     batch: Tuple[Float[Array, "n x"], Float[Array, "n y"]], 
     sharding: Optional[NamedSharding] = None
 ) -> Tuple[Float[Array, "n x"], Float[Array, "n y"]]:
+    """
+    Applies sharding to a batch of data for distributed processing.
+
+    Args:
+        batch (`Tuple[Array, Array]`): A tuple containing:
+            - Simulation data of shape `(n, x)`.
+            - Parameter data of shape `(n, y)`.
+        sharding (`jax.sharding.NamedSharding`): An optional `NamedSharding` 
+            object to define how the batch is distributed across devices.
+
+    Returns:
+        (`Tuple[Array, Array]`): The batch, potentially sharded according to the specified `sharding`.
+
+    Notes:
+        - If `sharding` is not provided, the batch remains unchanged.
+        - Useful for distributing data across devices in multi-GPU or TPU setups.
+    """
     if sharding:
         batch = eqx.filter_shard(batch, sharding)
     return batch
@@ -37,6 +54,23 @@ def apply_ema(
     model: eqx.Module, 
     ema_rate: float = 0.9999
 ) -> eqx.Module:
+    """
+    Updates an Exponential Moving Average (EMA) model based on the current model parameters.
+
+    Args:
+        ema_model: The current EMA model (`eqx.Module`).
+        model: The current model (`eqx.Module`) whose parameters are used to update the EMA model.
+        ema_rate: The decay rate for the EMA. Defaults to `0.9999`.
+
+    Returns:
+        The updated EMA model.
+
+    Notes:
+        - EMA is computed using the formula:
+          `p_ema = p_ema * ema_rate + p * (1 - ema_rate)`, 
+          where `p_ema` are the EMA parameters, and `p` are the model parameters.
+        - Inexact arrays are updated while preserving non-trainable state.
+    """
     ema_fn = lambda p_ema, p: p_ema * ema_rate + p * (1. - ema_rate)
     m_, _m = eqx.partition(model, eqx.is_inexact_array)
     e_, _e = eqx.partition(ema_model, eqx.is_inexact_array)
@@ -45,6 +79,21 @@ def apply_ema(
 
 
 def clip_grad_norm(grads: PyTree, max_norm: float) -> PyTree:
+    """
+    Clips the gradient norm of a PyTree of gradients to a specified maximum.
+
+    Args:
+        grads: A PyTree containing the gradients to be clipped.
+        max_norm: The maximum allowable norm for the gradients.
+
+    Returns:
+        A PyTree of gradients where the norm is clipped to `max_norm` if it exceeds the limit.
+
+    Notes:
+        - Uses L2 norm for clipping.
+        - Preserves the relative proportions of gradients while scaling them.
+        - Avoids division by zero by adding a small epsilon (`1e-6`) to the denominator.
+    """
     norm = jnp.linalg.norm(
         jax.tree.leaves(
             jax.tree.map(jnp.linalg.norm, grads)
@@ -67,6 +116,30 @@ def make_step(
     clip_max_norm: Optional[float] = None,
     replicated_sharding: Optional[PositionalSharding] = None,
 ) -> Tuple[eqx.Module, PyTree, Float[Array, ""]]:
+    """
+    Performs a single optimization step for a neural density estimator.
+
+    Args:
+        nde: The neural density estimator model (`eqx.Module`) being optimized.
+        x: The input data of shape `(b, x)`.
+        y: The target data of shape `(b, y)`.
+        opt_state: The optimizer state (`PyTree`) used to compute parameter updates.
+        opt: The optimizer object (`Optimiser`) for computing updates.
[…]
