### 2025-01-10 c05a37f
test fitting nn, docstrings
