### 2025-01-10 homerjed/sbiax#7 @homerjed
Thanks - good ideas - I forgot about these two...

I've changed them now, also removed `sbiax/inference/npe.npy` which would need to have docstrings if it were to be used.

Also added a test for fitting the compression neural networks.
