### 2025-01-10 486797f
clearing stuff up, docstrings
