+++ configs/__init__.py
@@ -4,18 +4,20 @@ from ml_collections import ConfigDict
 
 
 def save_config(config: ConfigDict, filepath: str):
+    """ Save a config to a yaml file. """
     with open(filepath, 'w') as f:
         yaml.dump(config.to_dict(), f)
 
 
 def load_config(filepath: str) -> ConfigDict:
+    """ Load a config to a yaml file. """
     with open(filepath, 'r') as f:
         config_dict = yaml.safe_load(f)
     return ConfigDict(config_dict)
 
 
-def make_dirs(results_dir):
-
+def make_dirs(results_dir: str) -> None:
+    """ Create directories for saving experimental results. """
     print("RESULTS_DIR:\n", results_dir)
     if not os.path.exists(results_dir):
         os.makedirs(results_dir, exist_ok=True)
@@ -24,18 +26,15 @@ def make_dirs(results_dir):
         "data/", "posteriors/", "models/", "figs/"
     ]
     for _dir in dirs:
-        os.makedirs(
-            os.path.join(results_dir, _dir), 
-            exist_ok=True
+        os.makedirs(os.path.join(results_dir, _dir), exist_ok=True)
+
+
+def get_results_dir(config: ConfigDict, base_dir="./") -> str:
+    """ Format the results directory based on parameters of a config. """
+    results_dir = os.path.join(
+        base_dir, "results/{}/{}/{}/".format(
+            config.sbi_type, config.exp_name, config.seed
         )
-
-
-def get_results_dir(config, base_dir="./"):
-    results_dir = (
-        "results/" +  
-        f"{config.sbi_type}/" + 
-        f"{config.exp_name}/" + 
-        f"{config.seed}/"
     )
     make_dirs(results_dir)
     return results_dir
