+++ README.md
@@ -90,6 +90,10 @@ Install via
 
 and have a look at [examples](https://github.com/homerjed/sbiax/tree/main/examples).
 
+### Contributing
+
+Want to add something? See `CONTRIBUTING.md`.
+
 ### Citation
 
 If you found this library to be useful in academic work, please cite:  <!--([arXiv link](https://arxiv.org/abs/2111.00254)) -->
