### 2025-01-13 homerjed/sbiax#1 @jbytecode
Due to the JOSS submission:

https://github.com/openjournals/joss-reviews/issues/7606
