### 2025-01-15 3fcb6a5
test assertions, test nuts mcmc, type fixes, paper ref
