+++ paper/paper.bib
@@ -151,7 +151,7 @@
       url={https://arxiv.org/abs/2202.02435}, 
 }
 
-@misc{homer2024simulationbasedinferencedodelsonschneidereffect,
+@misc{homer,
       title={Simulation-based inference has its own Dodelson-Schneider effect (but it knows that it does)}, 
       author={Jed Homer and Oliver Friedrich and Daniel Gruen},
       year={2024},
