### 2025-01-15 a3282e3
jnp.asarray bug... separate tests
