### 2025-01-15 ce618bb
rename tests, add paper ref to md/bib again...
