+++ tests/test_toy_example.py
@@ -143,6 +143,7 @@ def test_toy_nle():
         results_dir=results_dir
     )
 
+    # Check losses of NDEs
     assert all(
         [
             jnp.all(jnp.isfinite(jnp.asarray(stats[n]["train_losses"])))
@@ -202,205 +203,208 @@ def test_toy_nle():
     assert all([samples.shape[-1] == parameter_dim for samples, _ in posteriors])
 
 
-def test_toy_npe():
+# Test takes a long time so only run NLE for now
 
-    import os
-    import multiprocessing
+# def test_toy_npe():
 
-    os.environ["XLA_FLAGS"] = "--xla_force_host_platform_device_count={}".format(
-        multiprocessing.cpu_count()
-    )
+#     import os
+#     import multiprocessing
 
-    import jax
-    import jax.numpy as jnp
-    import jax.random as jr
-    import equinox as eqx
-    import diffrax as dfx
-    import optax
-    import tensorflow_probability.substrates.jax.distributions as tfd
+#     os.environ["XLA_FLAGS"] = "--xla_force_host_platform_device_count={}".format(
+#         multiprocessing.cpu_count()
+#     )
 
-    from configs import make_dirs
+#     import jax
+#     import jax.numpy as jnp
+#     import jax.random as jr
+#     import equinox as eqx
+#     import diffrax as dfx
+#     import optax
+#     import tensorflow_probability.substrates.jax.distributions as tfd
 
-    from sbiax.utils import make_df, get_shardings
-    from sbiax.ndes import Ensemble, MAF, CNF, Scaler
-    from sbiax.train import train_ensemble
-    from sbiax.inference import nuts_sample
-    from sbiax.compression.nn import fit_nn
+#     from configs import make_dirs
 
-    key = jr.key(0)
+#     from sbiax.utils import make_df, get_shardings
+#     from sbiax.ndes import Ensemble, MAF, CNF, Scaler
+#     from sbiax.train import train_ensemble
+#     from sbiax.inference import nuts_sample
+#     from sbiax.compression.nn import fit_nn
 
-    key, model_key, train_key, sample_key, data_key = jr.split(key, 5)
+#     key = jr.key(0)
 
-    results_dir = "results/toy/"
+#     key, model_key, train_key, sample_key, data_key = jr.split(key, 5)
 
-    make_dirs(results_dir)
+#     results_dir = "results/toy/"
 
-    n_sims = 10_000
-    data_dim = 100
-    parameter_dim = 2
+#     make_dirs(results_dir)
 
-    alpha = jnp.array([0.5, 0.5]) # True values of mu and sigma
-    lower = jnp.array([0., 0.1])
-    upper = jnp.array([1., 1.])
-    parameter_names = [r"$\mu$", r"$\sigma$"]
+#     n_sims = 10_000
+#     data_dim = 100
+#     parameter_dim = 2
 
-    parameter_prior = tfd.Blockwise(
-        [tfd.Uniform(lower[0], upper[0]), tfd.Uniform(lower[1], upper[1])]
-    )
+#     alpha = jnp.array([0.5, 0.5]) # True values of mu and sigma
+#     lower = jnp.array([0., 0.1])
+#     upper = jnp.array([1., 1.])
+#     parameter_names = [r"$\mu$", r"$\sigma$"]
 
-    def simulator(key, mu, sigma):
-        # Draw data d ~ G[d|mu, I * sigma]
-        return jr.multivariate_normal(key, jnp.ones(data_dim) * mu, jnp.eye(data_dim) * sigma)
+#     parameter_prior = tfd.Blockwise(
[…]
