### 2025-01-15 db92844
rename tests, add paper ref to md/bib again...
