+++ paper/paper.bib
@@ -160,7 +160,7 @@
       primaryClass={astro-ph.CO},
       url={https://arxiv.org/abs/2412.02311}, 
 }
-
+ 
 @article{sbimacke,
   doi = {10.21105/joss.02505},
   url = {https://doi.org/10.21105/joss.02505},
