### 2025-01-17 #7606 @jbytecode
@homerjed - Please perform a full check of the paper, it has some typos (datavector), language mistakes (sbiax is a code),  and missing references, e.g. `As shown in (?)`, and ping me when you've done with them. I will perform a full editorial check then.
