+++ paper/paper.bib
@@ -174,7 +174,7 @@
   journal = {Journal of Open Source Software}
 }
 
-@software{sbidirmeir,
+@software{sbidirmeier,
   author = {Simon Dirmeir},
   title = {SBIJAX: Simulation-based inference in JAX.},
   url = {https://github.com/dirmeier/sbijax},
