+++ paper/paper.bib
@@ -181,3 +181,43 @@
   version = {0.3.0},
   year = {2024},
 }
+
+@misc{Euclid,
+      title={Euclid Definition Study Report}, 
+      author={R. Laureijs and J. Amiaux and S. Arduini and J. -L. Auguères and J. Brinchmann and R. Cole and M. Cropper and C. Dabin and L. Duvet and A. Ealet and B. Garilli and P. Gondoin and L. Guzzo and J. Hoar and H. Hoekstra and R. Holmes and T. Kitching and T. Maciaszek and Y. Mellier and F. Pasian and W. Percival and J. Rhodes and G. Saavedra Criado and M. Sauvage and R. Scaramella and L. Valenziano and S. Warren and R. Bender and F. Castander and A. Cimatti and O. Le Fèvre and H. Kurki-Suonio and M. Levi and P. Lilje and G. Meylan and R. Nichol and K. Pedersen and V. Popa and R. Rebolo Lopez and H. -W. Rix and H. Rottgering and W. Zeilinger and F. Grupp and P. Hudelot and R. Massey and M. Meneghetti and L. Miller and S. Paltani and S. Paulin-Henriksson and S. Pires and C. Saxton and T. Schrabback and G. Seidel and J. Walsh and N. Aghanim and L. Amendola and J. Bartlett and C. Baccigalupi and J. -P. Beaulieu and K. Benabed and J. -G. Cuby and D. Elbaz and P. Fosalba and G. Gavazzi and A. Helmi and I. Hook and M. Irwin and J. -P. Kneib and M. Kunz and F. Mannucci and L. Moscardini and C. Tao and R. Teyssier and J. Weller and G. Zamorani and M. R. Zapatero Osorio and O. Boulade and J. J. Foumond and A. Di Giorgio and P. Guttridge and A. James and M. Kemp and J. Martignac and A. Spencer and D. Walton and T. Blümchen and C. Bonoli and F. Bortoletto and C. Cerna and L. Corcione and C. Fabron and K. Jahnke and S. Ligori and F. Madrid and L. Martin and G. Morgante and T. Pamplona and E. Prieto and M. Riva and R. Toledo and M. Trifoglio and F. Zerbi and F. Abdalla and M. Douspis and C. Grenet and S. Borgani and R. Bouwens and F. Courbin and J. -M. Delouis and P. Dubath and A. Fontana and M. Frailis and A. Grazian and J. Koppenhöfer and O. Mansutti and M. Melchior and M. Mignoli and J. Mohr and C. Neissner and K. Noddle and M. Poncet and M. Scodeggio and S. Serrano and N. Shane and J. -L. Starck and C. Surace and A. Taylor and G. Verdoes-Kleijn and C. Vuerli and O. R. Williams and A. Zacchei and B. Altieri and I. Escudero Sanz and R. Kohley and T. Oosterbroek and P. Astier and D. Bacon and S. Bardelli and C. Baugh and F. Bellagamba and C. Benoist and D. Bianchi and A. Biviano and E. Branchini and C. Carbone and V. Cardone and D. Clements and S. Colombi and C. Conselice and G. Cresci and N. Deacon and J. Dunlop and C. Fedeli and F. Fontanot and P. Franzetti and C. Giocoli and J. Garcia-Bellido and J. Gow and A. Heavens and P. Hewett and C. Heymans and A. Holland and Z. Huang and O. Ilbert and B. Joachimi and E. Jennins and E. Kerins and A. Kiessling and D. Ki
[…]
