+++ paper/paper.bib
@@ -221,3 +221,17 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2016MNRAS.460.1270D},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@article{ABC,
+author = {Donald B. Rubin},
+title = {{Bayesianly Justifiable and Relevant Frequency Calculations for the Applied Statistician}},
+volume = {12},
+journal = {The Annals of Statistics},
+number = {4},
+publisher = {Institute of Mathematical Statistics},
+pages = {1151 -- 1172},
+keywords = {62-07, Calibration, Empirical Bayes, inference, model monitoring, operating characteristics, posterior predictive checks, Stopping rules},
+year = {1984},
+doi = {10.1214/aos/1176346785},
+URL = {https://doi.org/10.1214/aos/1176346785}
+}
