+++ paper/paper.bib
@@ -192,16 +192,14 @@
       url={https://arxiv.org/abs/1110.3193}, 
 }
 
-@article{DESI,
-    author = "Levi, Michael E. and others",
-    collaboration = "DESI",
-    title = "{The Dark Energy Spectroscopic Instrument (DESI)}",
-    eprint = "1907.10688",
-    archivePrefix = "arXiv",
-    primaryClass = "astro-ph.IM",
-    reportNumber = "FERMILAB-PUB-19-434-AE",
-    month = "7",
-    year = "2019"
+@misc{DESI,
+      title={The Dark Energy Spectroscopic Instrument (DESI)}, 
+      author={Michael E. Levi and Lori E. Allen and Anand Raichoor and Charles Baltay and Segev BenZvi and Florian Beutler and Adam Bolton and Francisco J. Castander and Chia-Hsun Chuang and Andrew Cooper and Jean-Gabriel Cuby and Arjun Dey and Daniel Eisenstein and Xiaohui Fan and Brenna Flaugher and Carlos Frenk and Alma X. Gonzalez-Morales and Or Graur and Julien Guy and Salman Habib and Klaus Honscheid and Stephanie Juneau and Jean-Paul Kneib and Ofer Lahav and Dustin Lang and Alexie Leauthaud and Betta Lusso and Axel de la Macorra and Marc Manera and Paul Martini and Shude Mao and Jeffrey A. Newman and Nathalie Palanque-Delabrouille and Will J. Percival and Carlos Allende Prieto and Constance M. Rockosi and Vanina Ruhlmann-Kleider and David Schlegel and Hee-Jong Seo and Yong-Seon Song and Greg Tarle and Risa Wechsler and David Weinberg and Christophe Yeche and Ying Zu},
+      year={2019},
+      eprint={1907.10688},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.IM},
+      url={https://arxiv.org/abs/1907.10688}, 
 }
 
 @ARTICLE{DES,
@@ -234,4 +232,14 @@ keywords = {62-07, Calibration, Empirical Bayes, inference, model monitoring, op
 year = {1984},
 doi = {10.1214/aos/1176346785},
 URL = {https://doi.org/10.1214/aos/1176346785}
+}
+
+@misc{NRE,
+      title={Towards Reliable Simulation-Based Inference with Balanced Neural Ratio Estimation}, 
+      author={Arnaud Delaunoy and Joeri Hermans and François Rozet and Antoine Wehenkel and Gilles Louppe},
+      year={2022},
+      eprint={2208.13624},
+      archivePrefix={arXiv},
+      primaryClass={stat.ML},
+      url={https://arxiv.org/abs/2208.13624}, 
 }
