### 2025-01-17 9c13b1e
fix manuscript and bibliography
