### 2025-01-17 #7606 @jbytecode
@homerjed - I've just sent a pull request (https://github.com/homerjed/sbiax/pull/14)

This PR includes some minor fixes in both the manuscript and the bibliography.

Please review them. Tell me if you see anything wrong with my changes. If you are agree with them, please apply.

Thank you in advance.
