### 2025-01-17 #7606 @jbytecode
@homerjed - While I check the whole stuff one more time, you can perform the required tasks as listed in

https://github.com/openjournals/joss-reviews/issues/7606#issuecomment-2598131397

- Please create a version tag in the repository, e.g., `v0.0.9`
- Create a Zenodo archive using this tag
- The title, author list, author orcids should match those in the submission
- report here the DOI, Zenodo page url and the version that is used for registration.

Thank you in advance
