### 2025-01-17 #7606 @homerjed
- version tag is 0.0.52
- author names / ORCID IDs same as paper for authors
- DOI: 10.5281/zenodo.14679226
- Zenodo page URL: https://zenodo.org/records/14679226
