### 2025-01-17 #7606 @jbytecode
The Zenodo archive should include the latest image of the repository labeled with a revision tag. The current reported Zenodo archive holds the manuscript pdf, not the repository content. The aim of creating a registration in the Zenodo is to attach the freeze latest reviewed software with the manuscript. JOSS will host the published manuscript, already.
