### 2025-01-17 #7606 @jbytecode
Please have a read of the JOSS submission docs:

> Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/), get a DOI for the archive, and update the review issue thread with the version number and DOI.

located in

https://joss.readthedocs.io/en/latest/submitting.html#review-process

and please don't rush.
