### 2025-01-17 #7606 @homerjed
I've uploaded a version-tagged release `sbiax-0.0.54.zip` from

https://github.com/homerjed/sbiax/releases/tag/0.0.54

to

https://zenodo.org/records/14679376

which gives a DOI of

10.5281/zenodo.14679376

Hopefully that's correct. Thanks.
