### 2025-01-17 #7606 @jbytecode
@homerjed - The metadata should match those in the Zenodo archive. In Zenodo archive, orcid values of authors are missing. Please at them in the archive page and then ping me.

Do you think the paper title is correct? You may want to remove the last char (Dot `.`) in the title.  If so we have to fix three places

- the title in the manuscript
- the title in the zenodo
- the title here (I can change it)

What do you think?
