### 2025-01-17 9b0cf91
Update paper.md: odd grammar etc.
