### 2025-01-17 #7606 @homerjed
Thank you! I've changed tiny bits of grammar etc. e.g. consistent use of MCMC-sampler like an earlier change you added before.
