### 2025-01-17 #7606 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
