### 2025-01-17 #7606 @crvernon
:wave: @homerjed - I need you to address the following before I can move to accept this for publication:

In your archive:
- [ ] I noticed that the current version released in your GitHub repository is v0.0.55 but the version you have set here and in your Zenodo archive is 0.0.54.  Do you need to update this?  Or is 0.0.54 the target version you want represented in this paper?
- [ ] Please edit the license in your archive to match the one used in your GitHub repository.

Let me know once these have been addressed and I'll move forward with accepting your paper for publication.  Thanks.
