### 2025-01-17 #7606 @homerjed
Hi @crvernon,

The version 0.0.55 has no meaningful changes so the 0.0.54 version is what I think should be published.

I have changed the archived Zenodo version to adopt the MIT license in keeping with the code on the repo.

LMK if I've missed something.

Thank you.
